refactor(levels): send log on delete by service instead of route

+++ src/services/levels.ts
@@ -1,12 +1,14 @@
 import {Document, ObjectId} from 'mongodb'
 import {exec} from '../db'
-import {Level, NewLevel} from '../models'
+import {Level, NewLevel, Severity} from '../models'
 import {Config} from '../config'
 import {EmailsService} from './emails'
+import {LogsService} from './logs'
 import {Service} from './base'
 
 interface Dependencies {
   emails: EmailsService
+  logs: LogsService
 }
 
 export class LevelsService extends Service<Dependencies, Config['levels']> {
@@ -21,8 +23,19 @@ export class LevelsService extends Service<Dependencies, Config['levels']> {
   }
 
   public async deleteOne(id: string): Promise<void> {
+    const level = (await this.toArray({filter: {_id: new ObjectId(id)}}))[0]
+    if (!level) {
+      console.warn(`Level with id=${id} does not exists`)
+      return
+    }
+
     await exec<Level>('levels', async (collection) => {
       await collection.deleteOne({_id: new ObjectId(id)})
+    })
+
+    await this.dependencies.logs.insertOneFromWeb({
+      message: `Removed level "${level.value}" (${level.when.toLocaleString()})`,
+      severity: Severity.Info,
     })
   }
